requirements.txt
PySide6>=6.5.0,<7.0.0
opencv-python>=4.8.0,<5.0.0
ultralytics>=8.0.0,<9.0.0
requests>=2.31.0,<3.0.0
Pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0
onnxruntime>=1.15.0,<2.0.0
torch>=2.0.0,<3.0.0
onnxruntime-gpu>=1.15.0,<2.0.0
torch>=2.0.0,<3.0.0+cu118
pyqt6-plugins>=6.5.0
qt6-tools>=6.5.0